+++ src/app/components/forms/select-box/terra-select-box.component.ts
@@ -166,8 +166,9 @@ export class TerraSelectBoxComponent implements OnInit, OnChanges
 
     public get emptyValueSelected():boolean
     {
-        return StringHelper.isNullUndefinedOrEmpty(this._selectedValue.caption.toString()) &&
-               StringHelper.isNullUndefinedOrEmpty(this._selectedValue.icon);
+        return isNullOrUndefined(this._selectedValue) ||
+               (StringHelper.isNullUndefinedOrEmpty(this._selectedValue.caption.toString()) &&
+               StringHelper.isNullUndefinedOrEmpty(this._selectedValue.icon));
     }
 
     public get value():any